Add firewalld rules to block DNS53 and HTTP selectively
At first, we add rules for blocking all DNS53 and HTTP on port 80.
When the primaryconnection changes, we allow just the DNS53 to the
DHCP DNS servers and HTTP as required for captive portal clearing.

After NetworkManager tells us we've cleared the captive portal, we
clear all the allow rules, blocking all plaintext communication.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>